+++ api/Makefile
@@ -56,7 +56,7 @@ publish:
 .PHONY: dev
 dev: export ENABLE_VIRTUAL_SMOOTHIE := true
 dev:
-	$(python) opentrons/server/main.py -P 31950
+	$(python) opentrons/main.py -P 31950
 
 .PHONY: wheel
 wheel: clean
